a bunch of impovements, now everything seems to work correctly

+++ weather_notificator/src/main/java/com/tinkoff_lab/service/n/impl/EmailServiceImpl.java
@@ -2,10 +2,13 @@ package com.tinkoff_lab.service.n.impl;
 
 import com.tinkoff_lab.dao.jpa.CityRepository;
 import com.tinkoff_lab.dao.jpa.EmailRepository;
+import com.tinkoff_lab.dao.jpa.UserRepository;
+import com.tinkoff_lab.dto.EmailUserDTO;
 import com.tinkoff_lab.dto.n.CityDTO;
 import com.tinkoff_lab.dto.n.EmailDTO;
 import com.tinkoff_lab.entity.CityPK;
 import com.tinkoff_lab.entity.Email;
+import com.tinkoff_lab.exception.EmailVerificationException;
 import com.tinkoff_lab.mapper.CityMapper;
 import com.tinkoff_lab.mapper.EmailMapper;
 import com.tinkoff_lab.service.n.EmailService;
@@ -15,10 +18,12 @@ import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
 @Slf4j
+@Transactional
 @Service
 @RequiredArgsConstructor
 @FieldDefaults(makeFinal = true, level = AccessLevel.PRIVATE)
@@ -27,6 +32,7 @@ public class EmailServiceImpl implements EmailService {
     CityMapper cityMapper;
     EmailRepository emailRepository;
     CityRepository cityRepository;
+    UserRepository userRepository;
 
     @Override
     public List<EmailDTO> getEmails() {
@@ -38,8 +44,19 @@ public class EmailServiceImpl implements EmailService {
     }
 
     @Override
-    public List<CityDTO> getEmailCities(String email) {
+    public List<CityDTO> getEmailCities(String email, long chatId) {
         log.debug("Starting getEmailCities method.");
+        var e = emailRepository
+                .findById(email)
+                .orElseThrow(() -> {
+                    log.error("Email with id {} not found!", email);
+                    return new EntityNotFoundException("Email not found!");
+                });
+
+        if (e.getUser().getChatId() != chatId) {
+            throw new EmailVerificationException("Your email isn't verified!");
+        }
+
         var cities = emailRepository.getUserCities(email);
         log.info("Successfully retrieved cities.");
         log.debug("Leaving getEmailCities method.");
@@ -53,7 +70,7 @@ public class EmailServiceImpl implements EmailService {
                 .findById(id)
                 .orElseThrow(() -> {
                     log.error("Email with id {} not found!", id);
-                    return new EntityNotFoundException("Room not found!");
+                    return new EntityNotFoundException("Email not found!");
                 });
 
         log.info("Successfully retrieved email with id {}.", id);
@@ -62,7 +79,7 @@ public class EmailServiceImpl implements EmailService {
     }
 
     @Override
-    public void deleteEmail(String id) {
+    public void deleteEmail(String id, long chatId) {
         log.debug("Starting deleteEmail method with id {}.", id);
         var email = emailRepository
                 .findById(id)
@@ -70,6 +87,10 @@ public class EmailServiceImpl implements EmailService {
                     log.error("Attempt to delete email with non-existent id {}.", id);
                     return new EntityNotFoundException("Email not found!");
                 });
+
+        if (email.getUser().getChatId() != chatId && email.getIsVerified()) {
+            throw new EmailVerificationException("You're not verified in this email!");
+        }
 
         emailRepository.deleteCitiesFromEmail(email.getEmail());
         emailRepository.delete(email);
@@ -89,7 +110,7 @@ public class EmailServiceImpl implements EmailService {
 
         log.debug("Updating email fields with new values.");
         email.setName(emailDTO.email());
-        email.setVerified(emailDTO.isVerified());
+        email.setIsVerified(emailDTO.isVerified());
 
         var updatedEmail = mapper.entityToDto(emailRepository.save(email));
         log.info("Successfully updated email with id {}.", emailDTO.email());
@@ -101,6 +122,11 @@ public class EmailServiceImpl implements EmailService {
     public EmailDTO addEmail(EmailDTO emailDTO) {
         log.debug("Starting addEmail method.");
         Email email = mapper.dtoToEntity(emailDTO);
+        var em = emailRepository.findById(email.getEmail());
+        if (em.isPresent() && em.get().getIsVerified()) {
+            throw new EmailVerificationException("Email already used!");
+        }
+
         var createdEmail = mapper.entityToDto(emailRepository.save(email));
         log.info("Successfully created email with id {}.", emailDTO.email());
         log.debug("Leaving addEmail method.");
@@ -108,25 +134,34 @@ public class EmailServiceImpl implements EmailService {
     }
 
     @Override
-    public void verifyEmail(String code) {
+    public void verifyEmail(String code, long chatId) {
         var email = emailRepository
                 .findByCode(code)
                 .orElseThrow(() -> new EntityNotFoundException("Email not found."));
 
-        email.setVerified(true);
+        var user = userRepository
+                .findById(chatId)
+                .orElseThrow(() -> new EntityNotFoundException("USer not found."));
+
+        email.setUser(user);
+        email.setIsVerified(true);
         emailRepository.save(email);
     }
 
-    public EmailDTO addCitiesToEmail(String emailName, List<CityDTO> cityDTOS) {
+    public EmailDTO addCitiesToEmail(String email, long chatId, List<CityDTO> cityDTOS) {
         //log.debug("Starting addCityToEmail method for city {}.", cityDTOS.city());
-        var email = emailRepository
-                .findById(emailName)
+        var e = emailRepository
+                .findById(email)
                 .orElseThrow(() -> {
-                    log.error("Attempt to update emailName with non-existent id {}.", emailName);
+                    log.error("Attempt to update emailName with non-existent id {}.", email);
                     return new EntityNotFoundException("Email not found.");
                 });
 
-        for(CityDTO cityDTO: cityDTOS) {
+        if (e.getUser().getChatId() != chatId) {
+            throw new EmailVerificationException("Your email isn't verified!");
+        }
+
+        for (CityDTO cityDTO : cityDTOS) {
             var city = cityRepository
                     .findById(new CityPK(cityDTO.city(), cityDTO.country()))
                     .orElseGet(() -> {
@@ -134,26 +169,30 @@ public class EmailServiceImpl implements EmailService {
                         return cityRepository.save(newCity);
                     });
 
-            email.addCity(city);
-            log.info("Successfully added emailName {} to city {}.", emailName, cityDTO.city());
+            e.addCity(city);
+            log.info("Successfully added emailName {} to city {}.", email, cityDTO.city());
         }
 
-        emailRepository.save(email);
+        emailRepository.save(e);
         log.debug("Leaving addCityToEmail method.");
-        return mapper.entityToDto(email);
+        return mapper.entityToDto(e);
     }
 
     @Override
-    public void removeCitiesFromEmail(String emailName, List<CityDTO> cityDTOS){
+    public void removeCitiesFromEmail(String email, long chatId, List<CityDTO> cityDTOS) {
         //log.debug("Starting removeCityFromEmail method for city {}.", cityDTO.city());
-        var email = emailRepository
-                .findById(emailName)
+        var e = emailRepository
+                .findById(email)
                 .orElseThrow(() -> {
-                    log.error("Attempt to update email with non-existent id {}.", emailName);
+                    log.error("Attempt to update email with non-existent id {}.", email);
                     return new EntityNotFoundException("Email not found.");
                 });
 
-        for(CityDTO cityDTO: cityDTOS) {
+        if (e.getUser().getChatId() != chatId) {
+            throw new EmailVerificationException("Your email isn't verified!");
+        }
+
+        for (CityDTO cityDTO : cityDTOS) {
             var city = cityRepository
                     .findById(new CityPK(cityDTO.city(), cityDTO.country()))
                     .orElseThrow(() -> {
@@ -161,12 +200,16 @@ public class EmailServiceImpl implements EmailService {
                         return new EntityNotFoundException("City not found.");
                     });
 
-            email.removeCity(city);
+            if(!e.getCities().contains(city)){
+                throw new EntityNotFoundException("You're trying to delete city that is not in your list of cities!");
+            }
+
+            e.removeCity(city);
             //city.getEmails().remove(email);
-            log.info("Successfully added email {} to city {}.", emailName, cityDTO.city());
+            log.info("Successfully added email {} to city {}.", email, cityDTO.city());
         }
 
-        emailRepository.save(email);
+        emailRepository.save(e);
         log.debug("Leaving removeCityFromEmail method.");
     }
 }
